Revert "Refactor admin panel structure and styles (#27)"
This reverts commit 20fd29d1642c828e26f5af2e0cd73256ae4d78a4.
