--- src/components/adminPanel/Phone/style.module.css
@@ -1,17 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-  max-width: 400px;
-}
-
-.input {
-  padding: 10px;
-  margin-bottom: 20px;
-}
-
-@media (max-width: 600px) {
-  .container {
-    max-width: 100%;
-  }
-}
