--- src/components/adminPanel/index.js
@@ -1,5 +1,0 @@
-export { default as Header } from './Header';
-export { default as Auth } from './Auth';
-export { default as Register } from './Register';
-export { default as AdminLayout } from './AdminLayout';
-export { default as Loading } from './Loading/loading';
