+++ src/components/adminPannel/fullPost/index.jsx
