+++ src/components/adminPannel/fullPost/style.module.css
