+++ src/components/adminPannel/menu/images/Team.png
