+++ src/components/adminPannel/password/style.module.css
@@ -1,0 +1,4 @@
+.input {
+  padding: 10px;
+  margin-bottom: 20px;
+}
